QCamera2: Fix compile error caused by ARRAY_SIZE() comparison

+++ camera/QCamera2/HAL/QCamera2HWI.cpp
@@ -1485,7 +1485,7 @@ int QCamera2HardwareInterface::initCapabilities(uint32_t cameraId,
     ATRACE_CALL();
     int rc = NO_ERROR;
     QCameraHeapMemory *capabilityHeap = NULL;
-    int i, x;
+    size_t i, x;
 
     /* Allocate memory for capability buffer */
     capabilityHeap = new QCameraHeapMemory(QCAMERA_ION_USE_CACHE);
